Add SignalEventDefinition configuration
