+++ src/components/InspectorExtensionManager.js
@@ -18,7 +18,7 @@ function moveAdvancedAccordionToBottom(accordionA, accordionB) {
 
 function addInspectorItem(inspectorItems, config) {
   const nodeIndex = inspectorItems.findIndex(item => {
-    return (config.id && config.id === item.id) || (config.name && config.name === item.name);
+    return (config.id && config.id === item.id) || (config.config.name && config.config.name === item.config.name);
   });
   if (nodeIndex === -1) {
     inspectorItems.push(config);
