+++ src/components/nodes/intermediateSignalThrowEvent/index.js
@@ -2,9 +2,10 @@ import component from './intermediateSignalThrowEvent.vue';
 import merge from 'lodash/merge';
 import cloneDeep from 'lodash/cloneDeep';
 import intermediateEventConfig from '@/components/nodes/intermediateEvent';
-import omit from 'lodash/omit';
+import signalEventDefinition from '../signalEventDefinition';
 
 export default merge(cloneDeep(intermediateEventConfig), {
+  ...signalEventDefinition,
   id: 'processmaker-modeler-intermediate-signal-throw-event',
   component,
   control: false,
@@ -18,31 +19,6 @@ export default merge(cloneDeep(intermediateEventConfig), {
       ],
     });
   },
-  inspectorData(node) {
-    return Object.entries(node.definition).reduce((data, [key, value]) => {
-      if (key === 'eventDefinitions') {
-        data.signalName = value[0].get('signalRef').name;
-      } else {
-        data[key] = value;
-      }
-
-      return data;
-    }, {});
-  },
-  inspectorHandler(value, node, setNodeProp) {
-    for (const key in omit(value, ['$type', 'eventDefinitions', 'signalName'])) {
-      if (node.definition[key] === value[key]) {
-        continue;
-      }
-
-      setNodeProp(node, key, value[key]);
-    }
-
-    const signal = node.definition.get('eventDefinitions')[0].signalRef;
-    if (signal.name !== value.signalName) {
-      signal.name = value.signalName;
-    }
-  },
   inspectorConfig: [
     {
       items: [
@@ -52,9 +28,9 @@ export default merge(cloneDeep(intermediateEventConfig), {
             {
               component: 'FormInput',
               config: {
-                label: 'Signal Name',
-                name: 'signalName',
-                helper: 'Enter the signal name that is unique from all other elements in the diagram',
+                label: 'Signal Ref',
+                name: 'signalRef',
+                helper: 'Enter the signal reference that this element triggers',
               },
             },
           ],
